+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/buildsi/symbolator/tree/master) (0.0.x)
+ - Bug with saving result if no missing symbols (0.0.12)
  - Adding json export for generate and compare (0.0.11)
  - Initial creation of project (0.0.1)